Add Table Inspectors
The change 486482 was unable to be automatically merged with the current state of its repository. This is a rebase of the change.

Adding two kind of table modificators to VE:
1. fullwidth, by toggling one to-be-implemented css class
2. style, by offering a dropdown menu to select from existing styles

the fullwidth functionally expects the following css styles:
.tablefullwidth {
	width: 100%;
}
additionally to standard mediawiki/bluespicecalumma table styles.

Activated more codesniffing;
Added skipping if not in regular VE mode;
Removed SpecialPage stub.

ERM #7606
[3.1]

Change-Id: I628bfa6b13f251c2d40ffa9992254363db4d2266