Enable Pasting of Text with embedded Links
With this Change pasting of text with embedded links
into Visual Editor will change in behaviour:

On pasting plain text http and https links will no more be surrounded with
nowiki tags, but will be displayed as external links.
On pasting plain text every linebreak will cause a paragraph, so new lines and
double lines will work.
Using wgUrlProtocols for this.

For pasting from browser and/or office the contained links are kept

Cherrypick to REL1_31
ERM#14510

Change-Id: I7d400686a25d7280f46c54b76318441a98cfa219