Fix text preload in IE11 and VE
IE11 refuses to work with XML that has no root element. So the
preloaded text cannot be loaded. By wrapping the content in a tag,
this issue is fixed. Unfortunately, the div tag cannot be used,
as it will be translated into wikitext. With body tag, it works
fine as this tag is stripped by VE after loading the content.

NEEDS CHERRY-PICK to master and rel1_31_dev

Requires Ic3d75bd423aac3d26434e66e686a94dd3bc752f0

ERM#14887

Change-Id: If14b9588862de45e6b764d2de727f03c98dfbd4e