Prevent selection loss when changing table options
If we want to enable input fields to be freely editable, that means that as soon
as they get the focus, surface will lose it and we will lose our selection.
We can restore the selection once the user closes the dialog, but it leads to somewhat
strange UX, since table inspector would close during option editing and re-appear after
dialog is closed.
Options are to keep this behaviour or go back to non-editable input fields, where value
can only be set with +/- buttons in increment steps

[3.1]

Change-Id: I748b1b7ed2adf99ca8c1e6869a9ec32ab17bd292
ERM: #14506