Remove PlainText TransferHandler
Best I can say, this handler was performing operations on each char in pasted text,
and basically converted all to plain text, causing VE to freak-out and not recognize
any of the tags, surrounding all in <nowiki>.

I dont see any purpose of this transferHandler, as i dont see how and why we
would want to parse links from the plain text, as in plain texts links must be
just like: http://...., which is recognized by VE by default, and if they are not
written like this, we dont know how to parse any random format that can be coming in.

WIP until its tested by someone who knows the original problem for which this TH was implemented,
to make sure original problem is not back.

Needs CP to master and possibly REL1_31 (probably wont work)

Change-Id: I1bfa84616829b71a87bf5b6d23c3e19835a837db
ERM: #15977