Fix autosave prompt after new page creation
The autosave prompt is triggered on second edit of a newly created
article in the same tab. The reason is that in the local sessionStore,
the oldid is still stored with 0 (from page creation). For some obscure
reason, the sessionStore is not cleared on first save. This commit fixes
this by making sure all autosave data is cleared on save in any case.

ERM#13364

NEEDS CP to REL1_31 and master

Change-Id: I36990127f2a48e06d000a260c257b1a34037abe4