Change how line breaks are inserted
Previously, we were overriding the "BreakNode". This node handles <br> tags.
We cannot do that, no way to insert  (and persist) a <br> node directly.

We tried wrapping it in a <span>, but that was recognized by unintentional line breaks,
and messed the source code up.

Eventhough i hate this approarch, i decided to make a "custom" line break node,
just for the purpose of "explicitly" inserting a line break. Its a span with specific class.
It looks bad and kinda duplicates functionality, but we can guarantee no unintentional consequences

[4.2.x]

Other tickets on topic ERM30012, ERM30942
ERM 31604

Change-Id: I58ff154f2141396b7f3424740e52e35bf06147b9